Bloomberg Subscription Real-time Data
C# Implementation Bloomberg Subscription to real-time data via windows
forms in Visual Studio 2012. It begins a TCP/IP connection with the
local Bloomberg Terminal to start a subscription for a particular stock.
It will process the last trade price, last trade size, and calculate the
VWAP and total money flow. Lastly, it graphs all of this data onto
.NET's chart class.

NOTE: A Bloomberg terminal is required to retrieve the data for this
application.